% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search-utils.R
\name{binRate}
\alias{binRate}
\title{Locator of the proper bin (index) of a rate.}
\usage{
binRate(rate, bounds, sorted = FALSE, nonnegative = TRUE, leOne = TRUE)
}
\arguments{
\item{rate}{Rate for which to find bin placement.}

\item{sorted}{Whether the bounds have been sorted.}

\item{nonnegative}{Whether the rate is required to be nonnegative.}

\item{leOne}{Whether the rate is required to be no more than one.}

\item{endpoints}{Vector of bin bounds.}
}
\value{
Index of the bin into which the rate should be placed.
}
\description{
\code{binRate} takes a vector of values that it interprets as endpoints of 
bins, among which a rate is to be placed. It wraps \code{findInterval}
binary search as the strategy for determining the index of the "bin" into
}
\seealso{
\code{\link[base]{findInterval}}
}
