% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds-utils.R
\name{countListItems}
\alias{countListItems}
\title{Counter of items of certain type(s) from a composite collection.}
\usage{
countListItems(composite, singletonTypes)
}
\arguments{
\item{composite}{The composite object (e.g., list) from which to count items.}

\item{singletonTypes}{The collection of names of object types of interest 
(e.g., \code{c("data.table", "data.frame")}).}
}
\value{
The cumulative count of items of one of the types named in 
        \code{singletonTypes}.
}
\description{
\code{countItems} takes a composite object and a collection of types from 
of objects of interest and counts how many objects of those types occur 
in total in the composite collection.
}
