% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats-utils.R
\name{divKL}
\alias{divKL}
\title{Calculator of Kullack-Leibler divergence}
\usage{
divKL(pA, pB, base = 2)
}
\arguments{
\item{pA}{Vector of probabilities definining one distribution,}

\item{pB}{Vector of probabilities defining another distribution.}

\item{base}{The logarithm base to use. Common values are 2 for a value 
in bits and natural constant \code{e} for a value in nats.}
}
\value{
Kullback-Leibler divergence between the given distributions.
}
\description{
\code{divKL} accepts a pair of vectors of probability vectors that 
it interprets as defining a pair of discrete distributions and uses 
the values to compute Kullback-Leibler divergence between the distributions. 
The probability vectors must be normalized and be of equal length.
The logarithm base can be provided as an optional argument to \code{base}. 
The default \code{base} is 2 so that the unit is the bit.
}
