% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info-theory.R
\name{entropy}
\alias{entropy}
\title{Alias for \code{shannonEntropy}}
\usage{
entropy(observations, normalized = FALSE)
}
\description{
Alias for \code{shannonEntropy}
}
\seealso{
\code{\link{shannonEntropy}}
}
