% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs-utils.R
\name{expandPath}
\alias{expandPath}
\title{Path expansion utility}
\usage{
expandPath(path)
}
\arguments{
\item{path}{The filesystem path to expand.}
}
\value{
Input path with user/environment variables expanded.
}
\description{
\code{expandPath} takes a path and expands user/environment variables.
\code{NULL}, \code{NA}, and empty path inputs are returned as-is. For 
environment variables, the assumption is that each will be prefixed with 
a dollar-sign \code{$}.
}
\seealso{
Other paths: \code{\link{.envVarPath}},
  \code{\link{.filepath}}, \code{\link{.isDir}},
  \code{\link{makeFilePath}}, \code{\link{makedirs}}
}
