% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats-utils.R
\name{insertPseudocounts}
\alias{insertPseudocounts}
\title{Assurance that each value of a vector is nonempty.}
\usage{
insertPseudocounts(observations, pseudocount = 1)
}
\arguments{
\item{observations}{Vector of values to ensure nonzero.}

\item{pseudocount}{Value with which to replace zeros.}
}
\value{
Updated observations vector.
}
\description{
\code{insertPseudocounts} takes an observation vector and a 
\code{pseudocount}, replacing all zeros in \code{observations} with the
\code{pseudocount}.
}
