% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-utils.R
\name{locatePackage}
\alias{locatePackage}
\title{Finder of package installation source location candidates.}
\usage{
locatePackage(pkg_name, paths = NULL, env_vars = c("STAGE", "CODE",
  "CODEBASE"))
}
\arguments{
\item{pkg_name}{Name of the package of interest.}

\item{paths}{Collection of full filepaths, each of which should be 
considered for the presence of package \code{pkg_name}.}

\item{env_vars}{Names of environment variables, each of which should 
point to a filepath to consider for the presence of 
\code{pkg_name}.}
}
\value{
The collection of paths to extant candidates for installation of 
        \code{pkg_name}.
}
\description{
\code{locatePackage} uses a package name, path(s) to folder(s), and 
some environment variables to find all of the extant location candidates 
for the installation from source for package indicated by \code{name}.
}
\seealso{
Other packages: \code{\link{refreshPackage}}
}
