% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs-utils.R
\name{makedirs}
\alias{makedirs}
\title{Builder of folder(s)}
\usage{
makedirs(dirpath, permissions = "0777", forceChmod = FALSE)
}
\arguments{
\item{dirpath}{Full path to the folder to create.}

\item{permissions}{Permissions string.}

\item{forceChmod}{Whether to change the permissions to the given 
setting if the folder already exists.}
}
\description{
\code{makedirs} creates a folder and any intermediates needed, 
doing nothing if the folder already exists.
}
