% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs-utils.R
\name{makeFilePath}
\alias{makeFilePath}
\title{Builder of path to file within project's results folder.}
\usage{
makeFilePath(base, sampleName, subdir, extension, suffix = NULL)
}
\arguments{
\item{base}{Project configuration object (environment- or list-like), from
which the parent folder is derived. Alternatively, an actual 
path to a parent folder.}

\item{sampleName}{Name for the sample for which to create filepath.}

\item{subdir}{What to assume about the project's folder structure. If a 
Boolean value, this indicates whether the expectation is for 
the file to live within a results subfolder named with the 
\code{sampleName}. If a string, this indicates the actual
name of the subfolder within the project's results folder. 
If a function, it should accept \code{sampleName} and the 
derive a name for the appropriate subfolder that way.}

\item{suffix}{Text to place between the sample name and the extension.
The default is no suffix, which will also result from an 
argument that's empty, \code{NULL}, or \code{FALSE}.}
}
\value{
Path expected for file for the project defined by \code{base} and 
        the sample indicated by \code{sampleName}. The correspondence 
        with file type/format is captured by the \code{subdir} and 
        \code{extension} specifications.
}
\description{
\code{makeFilePath} uses information about file paths from a project 
configuration object, in conjunction with a sample name and how to 
deal with subfolder(s), to create a path to the output file that's 
known to the project for the sample indicated by \code{sampleName}, and 
that corresponds to a particular type/format (e.g., as generated by a 
particular pipeline).
}
\seealso{
Other paths: \code{\link{.envVarPath}},
  \code{\link{.filepath}}, \code{\link{.isDir}},
  \code{\link{expandPath}}, \code{\link{makedirs}}
}
