% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-utils.R
\name{nOfEach}
\alias{nOfEach}
\title{Sampler for the first observation of each unique value for a variable.}
\usage{
nOfEach(dt, column, n)
}
\arguments{
\item{dt:}{The \code{data.table} from which to sample.}

\item{column:}{Name of the column of interest.}

\item{n:}{Number of first-occurring observation(s) to grab for each 
unique value of \code{column}.}
}
\description{
\code{oneOfEach} identifies the unique values from a particular column 
of a data
}
