% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info-theory.R
\name{normalizedEntropy}
\alias{normalizedEntropy}
\title{Alias for \code{normalizedShannonEntropy}}
\usage{
normalizedEntropy(observations)
}
\description{
Alias for \code{normalizedShannonEntropy}
}
\seealso{
\code{\link{shannonEntropy}}, \code{\link{normalizedShannonEntropy}}
}
