% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info-theory.R
\name{normalizedShannonEntropy}
\alias{normalizedShannonEntropy}
\title{Calculator of normalized Shannon entropy for given observations vector.}
\usage{
normalizedShannonEntropy(observations)
}
\description{
Calculator of normalized Shannon entropy for given observations vector.
}
\seealso{
\code{\link{shannonEntropy}}
}
