% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-utils.R
\name{oneOfEach}
\alias{oneOfEach}
\title{Sampler of first observation of a particular table column.}
\usage{
oneOfEach(...)
}
\arguments{
\item{...:}{Arguments to \code{nOfEach}.}
}
\value{
Data table with first occurrence of each unique value from the 
        indicated column in the given table.
}
\description{
\code{oneOfEach} grabs the first observation for each unique value 
for a particular column within a table.
}
\seealso{
\code{\link{nOfEach}}
}
