% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-utils.R
\name{RefreshPackage}
\alias{RefreshPackage}
\title{Installer and loader of a package, either local or from github}
\usage{
RefreshPackage(packPath, useLocal = FALSE, name = NULL,
  nameFromUrl = NULL)
}
\arguments{
\item{packPath}{Path to the package to install and load.}

\item{useLocal}{Whether to use local source for the installation.}

\item{name}{Name for the package to load.}

\item{nameFromUrl}{Strategy with which to infer package name from URL.}
}
\description{
\code{RefreshPackage} takes a path to a file or folder, or a URL,  
determines whether it exists, and installs it. In order for local 
source to be used, the path given must exist locally and \code{useLocal} 
must be true. Once the package given by \code{packPath} is installed, 
this function attempts to load the package, using either a given name or 
an inferred one. If no \code{name} is given and the name inference 
strategy employed fails to yield a viable package name candidate, then 
the attempt to load the package is skipped.
}
