% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functools.R
\name{updateFoldRight}
\alias{updateFoldRight}
\title{Helper for creating function argument to rightward list folder.}
\usage{
updateFoldRight(result, newArg, f)
}
\arguments{
\item{result}{The growing collection of a \code{Reduce} operation 
in which the "reduction" is actually list formation.}

\item{newArg}{The new argument being processed.}

\item{f}{The function to apply to the new argument to generate a 
new element for \code{result}.}
}
\description{
Working directly with \code{base::Reduce} can be a little tricky, 
so this function aims to simplify using it. It's intended to be 
partially applied with something like:
\code{pryr::partial(updateFoldRight, f=identity)}
and then passed as the function argument to \code{foldRightList}.
}
\seealso{
\code{\link{foldRightList}} for the intended destination 
         of a partially-applied version of this, 
         \code{\link[base]{Reduce}} for what's wrapped.
}
