% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats-utils.R
\name{countBinnedObservations}
\alias{countBinnedObservations}
\title{Counter of number of observations that fall into each of a number of bins}
\usage{
countBinnedObservations(observations, binBounds, strictLowerBound = TRUE,
  pseudocount = 0)
}
\arguments{
\item{observations}{The observed values that come from a theoretically 
continuous distribution to discretize.}

\item{binBounds}{The vector of values that defines the (inclusive) left 
endpoint for each bin, that collectively will 
discretize the hypothetically continuous distribution 
of observations.}

\item{strictLowerBound}{Whether the lower bin bound is expected to be 
strict. That is, no observation should fall 
beneath it. If \code{TRUE} and this happens, 
it's an error. If \code{TRUE} but valid, then 
the returned counts distribution excludes the low 
("0") bin defined by \code{findInterval}.}

\item{pseudocount}{The count for any bin that contains no observations.}
}
\value{
Observation count for each bin.
}
\description{
\code{countBinnedObservations} takes an observations vector and a vector 
of values that define the (inclusive) left endpoints of a series of 
bins that will discretize the distribution of observations.
}
\seealso{
\code{\link[base]{findInterval}}
}
