% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs-utils.R
\name{ExpandPath}
\alias{ExpandPath}
\title{Path expansion utility}
\usage{
ExpandPath(path)
}
\arguments{
\item{path}{The filesystem path to expand.}
}
\value{
Input path with user/environment variables expanded.
}
\description{
\code{ExpandPath} takes a path and expands user/environment variables.
\code{NULL}, \code{NA}, and empty path inputs are returned as-is. For 
environment variables, the assumption is that each will be prefixed with 
a dollar-sign \code{$}.
}
